requirements.txt
fastapi
uvicorn
jinja2
tensorflow
numpy
pillow
pillow-heif
python-multipart